Fixed #12.
Had to move converters over to common to resolve a potential circular
dependency. Made the DigitalDisplayView support negative numbers.
