+++ MineSearch/DigitalDisplay.Wpf/Resources/Digits.xaml
@@ -15,6 +15,42 @@
         <Setter Property="Opacity" Value="{StaticResource InactiveOpacity}" />
     </Style>
 
+    <!-- Minus Sign (-) -->
+    <ControlTemplate x:Key="MinusSign">
+        <Viewbox Height="{TemplateBinding Height}" Width="{TemplateBinding Width}">
+            <Canvas Width="11" Height="21">
+                <!-- Top -->
+                <Polygon
+                    Points="1,0 9,0 7,2 3,2"
+                    Style="{StaticResource DigitInactiveStyle}" />
+                <!-- Top Left -->
+                <Polygon
+                    Points="0,1 0,9 2,7 2,3"
+                    Style="{StaticResource DigitInactiveStyle}" />
+                <!-- Bottom Left -->
+                <Polygon
+                    Points="0,11 0,19 2,17, 2,13"
+                    Style="{StaticResource DigitInactiveStyle}" />
+                <!-- Bottom-->
+                <Polygon
+                    Points="1,20 9,20 7,18, 3,18"
+                    Style="{StaticResource DigitInactiveStyle}" />
+                <!-- Bottom Right -->
+                <Polygon
+                    Points="10,19 10,11 8,13 8,17"
+                    Style="{StaticResource DigitInactiveStyle}" />
+                <!-- Top Right -->
+                <Polygon
+                    Points="10,9 10,1 8,3 8,7"
+                    Style="{StaticResource DigitInactiveStyle}" />
+                <!-- Middle -->
+                <Polygon
+                    Points="2,10 3,9 7,9 8,10 7,11 3,11 2,10"
+                    Style="{StaticResource DigitActiveStyle}" />
+            </Canvas>
+        </Viewbox>
+    </ControlTemplate>
+    
     <!-- 0 -->
     <ControlTemplate x:Key="Digit0">
         <Viewbox Height="{TemplateBinding Height}" Width="{TemplateBinding Width}">
