+++ MineSearch/MineSearch.Common.Test/ConverterTests/InverseBooleanToVisibilityConverterTest.cs
@@ -1,8 +1,8 @@
 ﻿using System.Windows;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using MineSearch.Wpf.Converters;
+using MineSearch.Common.Converters;
 
-namespace MineSearch.Wpf.Test.ConverterTests
+namespace MineSearch.Common.Test.ConverterTests
 {
     [TestClass]
     public class InverseBooleanToVisibilityConverterTest
