+++ MineSearch/MineSearch.Common/MineSearch.Common.csproj
@@ -30,16 +30,24 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Xaml" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
+    <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BasePointGenerator.cs" />
+    <Compile Include="Converters\AdjacentMineCountToColorConverter.cs" />
+    <Compile Include="Converters\AdjacentMineCountToStringConverter.cs" />
+    <Compile Include="Converters\ExplosionSourceToColorConverter.cs" />
+    <Compile Include="Converters\InverseBooleanToVisibilityConverter.cs" />
     <Compile Include="RandomPointGenerator.cs" />
     <Compile Include="IMatrix.cs" />
     <Compile Include="IPointGenerator.cs" />
