+++ MineSearch/MineSearch.Wpf/App.xaml
@@ -5,7 +5,7 @@
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="Resources/Style.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/DigitalDisplay.Wpf;component/Resources/Digits.xaml"></ResourceDictionary>
+                <ResourceDictionary Source="pack://application:,,,/DigitalDisplay.Wpf;component/Resources/Digits.xaml" />
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </Application.Resources>
