+++ MineSearch/MineSearch.Wpf/Views/UserControls/SafeCellView.xaml
@@ -3,14 +3,14 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:converters="clr-namespace:MineSearch.Wpf.Converters"
+             xmlns:commonConverters="clr-namespace:MineSearch.Common.Converters;assembly=MineSearch.Common"
              mc:Ignorable="d" 
              d:DesignHeight="26" d:DesignWidth="26">
     <UserControl.Resources>
         <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
-        <converters:InverseBooleanToVisibilityConverter x:Key="InverseBooleanToVisibilityConverter" />
-        <converters:AdjacentMineCountToStringConverter x:Key="AdjacentMineCountToStringConverter" />
-        <converters:AdjacentMineCountToColorConverter x:Key="AdjacentMineCountToColorConverter" />
+        <commonConverters:InverseBooleanToVisibilityConverter x:Key="InverseBooleanToVisibilityConverter" />
+        <commonConverters:AdjacentMineCountToStringConverter x:Key="AdjacentMineCountToStringConverter" />
+        <commonConverters:AdjacentMineCountToColorConverter x:Key="AdjacentMineCountToColorConverter" />
     </UserControl.Resources>
     <Grid>
         <!-- Covered state -->
